+++ src/main/java/io/github/dsheirer/dsp/filter/channelizer/ChannelCalculator.java
@@ -75,7 +75,7 @@ public class ChannelCalculator
      */
     private void updateChannelBandwidth()
     {
-        mChannelBandwidth = (double)mSampleRate / (double)mChannelCount;
+        mChannelBandwidth = mSampleRate / (double)mChannelCount;
     }
 
     /**
